requirements.txt
twine
wheel
docker
pyyaml
python-dotenv
click
requests
beautifultable
